imap: move over imapclient.FetchItem
